Adding functions: (AdikDrum::getNumSounds, getNumSteps, for using in UIApp class)

+++ src/consoleuiapp.h
@@ -17,7 +17,7 @@ public:
     void run() override;
     void close() override;
     void displayMessage(const std::string& message) override;
-    void displayGrid(const std::vector<std::vector<bool>>& grid, std::pair<int, int> cursor) override;
+    void displayGrid(const std::vector<std::vector<bool>>& grid, std::pair<size_t, size_t> cursor) override;
 
 private:
     AdikDrum& adikDrum_;
